The method `replaceOutboundUserPermissions` does not delete user permissions anymore but just replaces the whole collection of the user outbound permissions. The method `replaceOutboundUserPermissions` was moved from `UserPermissionCommandService` to `UserCommandService`. Tests adjusted.
